Prepare calendar component for release
