+++ build-tools/utils/files.js
@@ -36,7 +36,8 @@ function listPublicItems(baseDir) {
         elem !== 'interfaces.ts' &&
         elem !== 'test-utils' &&
         elem !== 'theming' &&
-        elem !== 'contexts'
+        elem !== 'contexts' &&
+        elem !== 'calendar'
     );
 }
