+++ src/__tests__/utils.ts
@@ -17,6 +17,7 @@ export function getAllComponents(): string[] {
         name !== 'test-utils' &&
         name !== 'theming' &&
         name !== 'contexts' &&
+        name !== 'calendar' &&
         !name.includes('.') &&
         !name.includes('LICENSE')
     );
