+++ src/calendar/grid/index.tsx
@@ -1,12 +1,12 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import React, { useMemo } from 'react';
-import styles from '../../styles.css.js';
+import styles from '../styles.css.js';
 import GridDay from './day';
 import { KeyCode } from '../../internal/keycode';
 import { addDays, addWeeks, isSameMonth } from 'date-fns';
 import { getCalendarMonth } from 'mnth';
-import { DayIndex } from '../index';
+import { DayIndex } from '../internal';
 import { MoveFocusHandler } from '../utils/move-focus-handler';
 import { DatePickerProps } from '../../date-picker/interfaces';
 import rotateDayIndexes from '../utils/rotate-day-indexes';
