+++ src/calendar/header/button/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { CancelableEventHandler } from '../../../internal/events';
 import { ButtonProps } from '../../../button/interfaces';
 import { InternalButton } from '../../../button/internal';
-import styles from '../../../styles.css.js';
+import styles from '../../styles.css.js';
 
 interface HeaderButtonProps {
   ariaLabel: string;
