+++ src/calendar/index.tsx
@@ -2,11 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
+import useBaseComponent from '../internal/hooks/use-base-component';
+import { applyDisplayName } from '../internal/utils/apply-display-name';
 import { CalendarProps } from './interfaces';
 import InternalCalendar from './internal';
 
 export { CalendarProps };
 
 export default function Calendar({ locale = '', isDateEnabled = () => true, ...props }: CalendarProps) {
-  return <InternalCalendar {...props} locale={locale} isDateEnabled={isDateEnabled} />;
+  const baseComponentProps = useBaseComponent('Calendar');
+  return <InternalCalendar {...props} {...baseComponentProps} locale={locale} isDateEnabled={isDateEnabled} />;
 }
+
+applyDisplayName(Calendar, 'Calendar');
