+++ src/calendar/internal.tsx
@@ -4,7 +4,6 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { addDays, addMonths, getDaysInMonth, isSameMonth, startOfMonth } from 'date-fns';
 import styles from './styles.css.js';
-import { BaseComponentProps } from '../internal/base-component';
 import useFocusVisible from '../internal/hooks/focus-visible/index.js';
 import CalendarHeader from './header';
 import Grid from './grid';
@@ -14,23 +13,12 @@ import { memoizedDate } from './utils/memoized-date.js';
 import { useEffectOnUpdate } from '../internal/hooks/use-effect-on-update.js';
 import { normalizeLocale, normalizeStartOfWeek } from './utils/locales.js';
 import { formatDate } from '../internal/utils/date-time';
-import { fireNonCancelableEvent, NonCancelableEventHandler } from '../internal/events/index.js';
-import { DatePickerProps } from '../date-picker/interfaces.js';
+import { fireNonCancelableEvent } from '../internal/events/index.js';
 import checkControlled from '../internal/hooks/check-controlled/index.js';
 import clsx from 'clsx';
+import { CalendarProps } from './interfaces.js';
 
 export type DayIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6;
-
-interface CalendarProps extends BaseComponentProps {
-  value: string;
-  onChange?: NonCancelableEventHandler<DatePickerProps.ChangeDetail>;
-  locale?: string;
-  startOfWeek?: number;
-  isDateEnabled?: (date: Date) => boolean;
-  todayAriaLabel: string;
-  nextMonthAriaLabel: string;
-  previousMonthAriaLabel: string;
-}
 
 export default function Calendar({
   value,
