+++ src/date-picker/interfaces.ts
@@ -4,63 +4,13 @@ import { BaseComponentProps } from '../internal/base-component';
 import { FormFieldValidationControlProps } from '../internal/context/form-field-context';
 import { ExpandToViewport } from '../internal/components/dropdown/interfaces';
 import { NonCancelableEventHandler } from '../internal/events';
-
-export interface DatePickerBaseProps {
-  /**
-   * The current input value, in YYYY-MM-DD format.
-   */
-  value: string;
-
-  /**
-   * Defines whether a particular date is enabled in the calendar or not.
-   * If you disable a date in the calendar, users can still enter this date using a keyboard.
-   * We recommend that you also validate these constraints on the client-side and server-side
-   * as you would for other form elements.
-   * @param date
-   */
-  isDateEnabled?: DatePickerProps.IsDateEnabledFunction;
-
-  /**
-   * Specifies the locale to use to render month names and determine the starting day of the week.
-   * If you don't provide this, the locale is determined by the page and browser locales.
-   * Supported values and formats are listed in the
-   * [JavaScript Intl API specification](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl#Locale_identification_and_negotiation).
-   */
-  locale?: string;
-
-  /**
-   * Determines the starting day of the week. The values 0-6 map to Sunday-Saturday.
-   * By default the starting day of the week is defined by the locale, but you can use this property to override it.
-   */
-  startOfWeek?: number;
-
-  /**
-   * Used as part of the `aria-label` for today's date in the calendar.
-   */
-  todayAriaLabel: string;
-
-  /**
-   * Specifies an `aria-label` for the 'next month' button.
-   */
-  nextMonthAriaLabel: string;
-
-  /**
-   * Specifies an `aria-label` for the 'previous month' button.
-   */
-  previousMonthAriaLabel: string;
-
-  /**
-   * Called whenever a user changes the input value (by typing, pasting, or selecting a value).
-   * The event `detail` contains the current value of the field.
-   */
-  onChange?: NonCancelableEventHandler<DatePickerProps.ChangeDetail>;
-}
+import { CalendarProps } from '../calendar/interfaces';
 
 export interface DatePickerProps
   extends BaseComponentProps,
     FormFieldValidationControlProps,
     ExpandToViewport,
-    DatePickerBaseProps {
+    CalendarProps {
   /**
    * Specifies the placeholder text rendered when the value is an empty string.
    */
