+++ src/test-utils/dom/date-picker/index.ts
@@ -1,8 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { ComponentWrapper, ElementWrapper, usesDom, createWrapper } from '@cloudscape-design/test-utils-core/dom';
+
+import { ElementWrapper, usesDom, createWrapper } from '@cloudscape-design/test-utils-core/dom';
 import styles from '../../../date-picker/styles.selectors.js';
+import calendarStyles from '../../../calendar/styles.selectors.js';
 import ButtonWrapper from '../button';
+import CalendarWrapper from '../calendar';
 import BaseInputWrapper from '../input/base-input';
 import DropdownWrapper from '../internal/dropdown';
 
@@ -15,7 +18,7 @@ export default class DatePickerWrapper extends BaseInputWrapper {
    */
   findCalendar(options = { expandToViewport: false }): CalendarWrapper | null {
     const wrapper = options.expandToViewport ? createWrapper() : this;
-    return wrapper.findComponent(`.${styles.calendar}`, CalendarWrapper);
+    return wrapper.findComponent(`.${calendarStyles.root}`, CalendarWrapper);
   }
 
   findCalendarDropdown(): ElementWrapper | null {
@@ -36,33 +39,5 @@ export default class DatePickerWrapper extends BaseInputWrapper {
    */
   @usesDom setInputValue(value: string): void {
     return super.setInputValue(value);
-  }
-}
-
-export class CalendarWrapper extends ComponentWrapper {
-  /**
-   * Returns a day container on the calendar.
-   *
-   * @param row 1-based row index of the day.
-   * @param column 1-based column index of the day.
-   */
-  findDateAt(row: number, column: number): ElementWrapper {
-    return this.find(`.${styles['calendar-week']}:nth-child(${row}) .${styles['calendar-day']}:nth-child(${column})`)!;
-  }
-
-  findHeader(): ElementWrapper {
-    return this.findByClassName(styles['calendar-header'])!;
-  }
-
-  findPreviousMonthButton(): ButtonWrapper {
-    return this.findComponent(`.${styles['calendar-prev-month-btn']}`, ButtonWrapper)!;
-  }
-
-  findNextMonthButton(): ButtonWrapper {
-    return this.findComponent(`.${styles['calendar-next-month-btn']}`, ButtonWrapper)!;
-  }
-
-  findSelectedDate(): ElementWrapper {
-    return this.find(`.${styles['calendar-day-selected']}`)!;
   }
 }
